App/holistic_ai_app/backend/requirements.txt
Flask
psycopg2-binary
requests
python-dotenv
Flask-CORS
waitress 